feat: 적 자동 생성 실습 과제 1 구현 완료적과의 충돌 로직 추가
- 적과의 충돌 시 데미지 받는 로직 구현

+++ Assets/02.Scripts/Bullet/Bullet.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class Bullet : MonoBehaviour
 {
     public float MoveSpeed;
-    public float BulletDamage;
+    public int BulletDamage;
     private float _maxPositionX = 5.2f;
 
     private void Update()
